requirements-test.txt
# Requirements for shellai testing
pytest>=7.0.0
pyyaml>=6.0
torch>=1.9.0
transformers>=4.20.0

# Optional testing dependencies
pytest-cov>=4.0.0  # For coverage reports
pytest-timeout>=2.1.0  # For test timeouts
pytest-mock>=3.10.0  # For advanced mocking
